fixup! sns-testing: make used NNS neuron ID configurable
